Remove the subcommands of the dump command The POs now decided to end the support for subcommands
